1.refactor email and sns notice user event architecture
