+++ openeulerSever/src/main/resources/docs/RejectNoticeEmailTemplate.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<body style="font-family: Arial, sans-serif; line-height: 1.6; color: #333; margin: 0; padding: 0;">
+<div style="width: 100%; max-width: 600px; margin: 0 auto; padding: 20px; background-color: #f9f9f9; border: 1px solid #ddd;">
+    <div style="padding: 20px">
+        <p>尊敬的${userName}，</p>
+        <p>您的“${productName}”测评申请未能通过审核，驳回原因：${reason}</p>
+        <p>请及时登录测评系统处理！</p>
+    </div>
+</div>
+</body>
+</html>
